requirements.txt
aniso8601==8.1.0
appdirs==1.4.3
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pyrsistent==0.17.3
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.5
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==0.16.0
